@@ -1,16 +1,19 @@
 module moqueries.org/cli
 
-go 1.15
+go 1.19
 
 require (
-	github.com/dave/dst v0.26.2
+	github.com/dave/dst v0.27.2
 	github.com/spf13/cobra v1.6.1
+	golang.org/x/mod v0.8.0
 	golang.org/x/text v0.7.0
 	golang.org/x/tools v0.6.0
+	moqueries.org/runtime v0.2.1-0.20230514231133-79eabd1bc852
 )
 
 require (
-	github.com/stretchr/testify v1.8.0 // indirect
-	golang.org/x/mod v0.8.0
-	moqueries.org/runtime v0.2.0
+	github.com/inconshreveable/mousetrap v1.0.1 // indirect
+	github.com/spf13/pflag v1.0.5 // indirect
+	golang.org/x/sys v0.5.0 // indirect
+	moqueries.org/deephash v0.26.0 // indirect
 )
