@@ -4,10 +4,12 @@ package testmoqs
 
 import "io"
 
+// TODO: Need to have a full copy of this file in a test package?
+
 // NB: Keep in sync with ../generator_test.go TestGenerating
 
-//go:generate moqueries --destination moq_testmoqs_test.go UsualFn NoNamesFn NoResultsFn NoParamsFn NothingFn VariadicFn RepeatedIdsFn TimesFn DifficultParamNamesFn DifficultResultNamesFn PassByReferenceFn InterfaceParamFn InterfaceResultFn Usual
-//go:generate moqueries --destination exported/moq_exported_testmoqs.go --export UsualFn NoNamesFn NoResultsFn NoParamsFn NothingFn VariadicFn RepeatedIdsFn TimesFn DifficultParamNamesFn DifficultResultNamesFn PassByReferenceFn InterfaceParamFn InterfaceResultFn Usual
+//go:generate moqueries --destination moq_testmoqs_test.go UsualFn NoNamesFn NoResultsFn NoParamsFn NothingFn VariadicFn RepeatedIdsFn TimesFn DifficultParamNamesFn DifficultResultNamesFn PassByReferenceFn InterfaceParamFn InterfaceResultFn GenericParamsFn PartialGenericParamsFn GenericResultsFn PartialGenericResultsFn GenericInterfaceParamFn GenericInterfaceResultFn Usual GenericParams PartialGenericParams GenericResults PartialGenericResults GenericInterfaceParam GenericInterfaceResult
+//go:generate moqueries --destination exported/moq_exported_testmoqs.go --export UsualFn NoNamesFn NoResultsFn NoParamsFn NothingFn VariadicFn RepeatedIdsFn TimesFn DifficultParamNamesFn DifficultResultNamesFn PassByReferenceFn InterfaceParamFn InterfaceResultFn GenericParamsFn PartialGenericParamsFn GenericResultsFn PartialGenericResultsFn GenericInterfaceParamFn GenericInterfaceResultFn Usual GenericParams PartialGenericParams GenericResults PartialGenericResults GenericInterfaceParam GenericInterfaceResult
 
 // UsualFn is a typical function type
 type UsualFn func(sParam string, bParam bool) (sResult string, err error)
@@ -69,12 +71,34 @@ type InterfaceResultReader struct {
 	Err     error
 }
 
-func (r *InterfaceResultReader) Read(p []byte) (int, error) {
+func (r *InterfaceResultReader) Read([]byte) (int, error) {
 	return 0, nil
 }
 
 // InterfaceResultFn tests returning interface results
 type InterfaceResultFn func(sParam string, bParam bool) (r io.Reader)
+
+// GenericParamsFn has all generic parameters
+type GenericParamsFn[S, B any] func(S, B) (string, error)
+
+// PartialGenericParamsFn has some generic parameters
+type PartialGenericParamsFn[S any] func(S, bool) (string, error)
+
+// GenericResultsFn has all generic results
+type GenericResultsFn[S ~string, E error] func(string, bool) (S, E)
+
+// PartialGenericResultsFn has some generic results
+type PartialGenericResultsFn[S ~string] func(string, bool) (S, error)
+
+type MyWriter io.Writer
+
+// GenericInterfaceParamFn tests passing generic interface parameters
+type GenericInterfaceParamFn[W MyWriter] func(w W) (sResult string, err error)
+
+type MyReader io.Reader
+
+// GenericInterfaceResultFn tests returning generic interface results
+type GenericInterfaceResultFn[R MyReader] func(sParam string, bParam bool) (r R)
 
 // Usual combines all the above function types into an interface
 //
@@ -94,4 +118,28 @@ type Usual interface {
 	InterfaceParam(w io.Writer) (sResult string, err error)
 	InterfaceResult(sParam string, bParam bool) (r io.Reader)
 	FnParam(fn func())
+}
+
+type GenericParams[S, B any] interface {
+	Usual(S, B) (string, error)
+}
+
+type PartialGenericParams[S any] interface {
+	Usual(S, bool) (string, error)
+}
+
+type GenericResults[S ~string, E error] interface {
+	Usual(string, bool) (S, E)
+}
+
+type PartialGenericResults[S ~string] interface {
+	Usual(string, bool) (S, error)
+}
+
+type GenericInterfaceParam[W MyWriter] interface {
+	Usual(w W) (sResult string, err error)
+}
+
+type GenericInterfaceResult[R MyReader] interface {
+	Usual(sParam string, bParam bool) (r R)
 }
