@@ -45,10 +45,12 @@ type moqTypeCache struct {
 				testImport moq.ParamIndexing
 			}
 			IsComparable struct {
-				expr moq.ParamIndexing
+				expr       moq.ParamIndexing
+				parentType moq.ParamIndexing
 			}
 			IsDefaultComparable struct {
-				expr moq.ParamIndexing
+				expr       moq.ParamIndexing
+				parentType moq.ParamIndexing
 			}
 			FindPackage struct {
 				dir moq.ParamIndexing
@@ -250,13 +252,22 @@ type moqTypeCache_Type_anyParams struct {
 }
 
 // moqTypeCache_IsComparable_params holds the params of the TypeCache type
-type moqTypeCache_IsComparable_params struct{ expr dst.Expr }
+type moqTypeCache_IsComparable_params struct {
+	expr       dst.Expr
+	parentType ast.TypeInfo
+}
 
 // moqTypeCache_IsComparable_paramsKey holds the map key params of the
 // TypeCache type
 type moqTypeCache_IsComparable_paramsKey struct {
-	params struct{ expr dst.Expr }
-	hashes struct{ expr hash.Hash }
+	params struct {
+		expr       dst.Expr
+		parentType ast.TypeInfo
+	}
+	hashes struct {
+		expr       hash.Hash
+		parentType hash.Hash
+	}
 }
 
 // moqTypeCache_IsComparable_resultsByParams contains the results for a given
@@ -269,11 +280,11 @@ type moqTypeCache_IsComparable_resultsByParams struct {
 
 // moqTypeCache_IsComparable_doFn defines the type of function needed when
 // calling andDo for the TypeCache type
-type moqTypeCache_IsComparable_doFn func(expr dst.Expr)
+type moqTypeCache_IsComparable_doFn func(expr dst.Expr, parentType ast.TypeInfo)
 
 // moqTypeCache_IsComparable_doReturnFn defines the type of function needed
 // when calling doReturnResults for the TypeCache type
-type moqTypeCache_IsComparable_doReturnFn func(expr dst.Expr) (bool, error)
+type moqTypeCache_IsComparable_doReturnFn func(expr dst.Expr, parentType ast.TypeInfo) (bool, error)
 
 // moqTypeCache_IsComparable_results holds the results of the TypeCache type
 type moqTypeCache_IsComparable_results struct {
@@ -309,13 +320,22 @@ type moqTypeCache_IsComparable_anyParams struct {
 
 // moqTypeCache_IsDefaultComparable_params holds the params of the TypeCache
 // type
-type moqTypeCache_IsDefaultComparable_params struct{ expr dst.Expr }
+type moqTypeCache_IsDefaultComparable_params struct {
+	expr       dst.Expr
+	parentType ast.TypeInfo
+}
 
 // moqTypeCache_IsDefaultComparable_paramsKey holds the map key params of the
 // TypeCache type
 type moqTypeCache_IsDefaultComparable_paramsKey struct {
-	params struct{ expr dst.Expr }
-	hashes struct{ expr hash.Hash }
+	params struct {
+		expr       dst.Expr
+		parentType ast.TypeInfo
+	}
+	hashes struct {
+		expr       hash.Hash
+		parentType hash.Hash
+	}
 }
 
 // moqTypeCache_IsDefaultComparable_resultsByParams contains the results for a
@@ -328,11 +348,11 @@ type moqTypeCache_IsDefaultComparable_resultsByParams struct {
 
 // moqTypeCache_IsDefaultComparable_doFn defines the type of function needed
 // when calling andDo for the TypeCache type
-type moqTypeCache_IsDefaultComparable_doFn func(expr dst.Expr)
+type moqTypeCache_IsDefaultComparable_doFn func(expr dst.Expr, parentType ast.TypeInfo)
 
 // moqTypeCache_IsDefaultComparable_doReturnFn defines the type of function
 // needed when calling doReturnResults for the TypeCache type
-type moqTypeCache_IsDefaultComparable_doReturnFn func(expr dst.Expr) (bool, error)
+type moqTypeCache_IsDefaultComparable_doReturnFn func(expr dst.Expr, parentType ast.TypeInfo) (bool, error)
 
 // moqTypeCache_IsDefaultComparable_results holds the results of the TypeCache
 // type
@@ -449,10 +469,12 @@ func newMoqTypeCache(scene *moq.Scene, config *moq.Config) *moqTypeCache {
 					testImport moq.ParamIndexing
 				}
 				IsComparable struct {
-					expr moq.ParamIndexing
+					expr       moq.ParamIndexing
+					parentType moq.ParamIndexing
 				}
 				IsDefaultComparable struct {
-					expr moq.ParamIndexing
+					expr       moq.ParamIndexing
+					parentType moq.ParamIndexing
 				}
 				FindPackage struct {
 					dir moq.ParamIndexing
@@ -471,10 +493,12 @@ func newMoqTypeCache(scene *moq.Scene, config *moq.Config) *moqTypeCache {
 				testImport moq.ParamIndexing
 			}
 			IsComparable struct {
-				expr moq.ParamIndexing
+				expr       moq.ParamIndexing
+				parentType moq.ParamIndexing
 			}
 			IsDefaultComparable struct {
-				expr moq.ParamIndexing
+				expr       moq.ParamIndexing
+				parentType moq.ParamIndexing
 			}
 			FindPackage struct {
 				dir moq.ParamIndexing
@@ -500,14 +524,18 @@ func newMoqTypeCache(scene *moq.Scene, config *moq.Config) *moqTypeCache {
 				testImport: moq.ParamIndexByValue,
 			},
 			IsComparable: struct {
-				expr moq.ParamIndexing
+				expr       moq.ParamIndexing
+				parentType moq.ParamIndexing
 			}{
-				expr: moq.ParamIndexByHash,
+				expr:       moq.ParamIndexByHash,
+				parentType: moq.ParamIndexByHash,
 			},
 			IsDefaultComparable: struct {
-				expr moq.ParamIndexing
+				expr       moq.ParamIndexing
+				parentType moq.ParamIndexing
 			}{
-				expr: moq.ParamIndexByHash,
+				expr:       moq.ParamIndexByHash,
+				parentType: moq.ParamIndexByHash,
 			},
 			FindPackage: struct {
 				dir moq.ParamIndexing
@@ -687,10 +715,11 @@ func (m *moqTypeCache_mock) Type(id dst.Ident, contextPkg string, testImport boo
 	return
 }
 
-func (m *moqTypeCache_mock) IsComparable(expr dst.Expr) (result1 bool, result2 error) {
+func (m *moqTypeCache_mock) IsComparable(expr dst.Expr, parentType ast.TypeInfo) (result1 bool, result2 error) {
 	m.moq.scene.T.Helper()
 	params := moqTypeCache_IsComparable_params{
-		expr: expr,
+		expr:       expr,
+		parentType: parentType,
 	}
 	var results *moqTypeCache_IsComparable_results
 	for _, resultsByParams := range m.moq.resultsByParams_IsComparable {
@@ -728,7 +757,7 @@ func (m *moqTypeCache_mock) IsComparable(expr dst.Expr) (result1 bool, result2 e
 	}
 
 	if result.doFn != nil {
-		result.doFn(expr)
+		result.doFn(expr, parentType)
 	}
 
 	if result.values != nil {
@@ -736,15 +765,16 @@ func (m *moqTypeCache_mock) IsComparable(expr dst.Expr) (result1 bool, result2 e
 		result2 = result.values.result2
 	}
 	if result.doReturnFn != nil {
-		result1, result2 = result.doReturnFn(expr)
+		result1, result2 = result.doReturnFn(expr, parentType)
 	}
 	return
 }
 
-func (m *moqTypeCache_mock) IsDefaultComparable(expr dst.Expr) (result1 bool, result2 error) {
+func (m *moqTypeCache_mock) IsDefaultComparable(expr dst.Expr, parentType ast.TypeInfo) (result1 bool, result2 error) {
 	m.moq.scene.T.Helper()
 	params := moqTypeCache_IsDefaultComparable_params{
-		expr: expr,
+		expr:       expr,
+		parentType: parentType,
 	}
 	var results *moqTypeCache_IsDefaultComparable_results
 	for _, resultsByParams := range m.moq.resultsByParams_IsDefaultComparable {
@@ -782,7 +812,7 @@ func (m *moqTypeCache_mock) IsDefaultComparable(expr dst.Expr) (result1 bool, re
 	}
 
 	if result.doFn != nil {
-		result.doFn(expr)
+		result.doFn(expr, parentType)
 	}
 
 	if result.values != nil {
@@ -790,7 +820,7 @@ func (m *moqTypeCache_mock) IsDefaultComparable(expr dst.Expr) (result1 bool, re
 		result2 = result.values.result2
 	}
 	if result.doReturnFn != nil {
-		result1, result2 = result.doReturnFn(expr)
+		result1, result2 = result.doReturnFn(expr, parentType)
 	}
 	return
 }
@@ -1508,10 +1538,11 @@ func (m *moqTypeCache) paramsKey_Type(params moqTypeCache_Type_params, anyParams
 	}
 }
 
-func (m *moqTypeCache_recorder) IsComparable(expr dst.Expr) *moqTypeCache_IsComparable_fnRecorder {
+func (m *moqTypeCache_recorder) IsComparable(expr dst.Expr, parentType ast.TypeInfo) *moqTypeCache_IsComparable_fnRecorder {
 	return &moqTypeCache_IsComparable_fnRecorder{
 		params: moqTypeCache_IsComparable_params{
-			expr: expr,
+			expr:       expr,
+			parentType: parentType,
 		},
 		sequence: m.moq.config.Sequence == moq.SeqDefaultOn,
 		moq:      m.moq,
@@ -1529,6 +1560,11 @@ func (r *moqTypeCache_IsComparable_fnRecorder) any() *moqTypeCache_IsComparable_
 
 func (a *moqTypeCache_IsComparable_anyParams) expr() *moqTypeCache_IsComparable_fnRecorder {
 	a.recorder.anyParams |= 1 << 0
+	return a.recorder
+}
+
+func (a *moqTypeCache_IsComparable_anyParams) parentType() *moqTypeCache_IsComparable_fnRecorder {
+	a.recorder.anyParams |= 1 << 1
 	return a.recorder
 }
 
@@ -1692,7 +1728,7 @@ func (r *moqTypeCache_IsComparable_fnRecorder) repeat(repeaters ...moq.Repeater)
 }
 
 func (m *moqTypeCache) prettyParams_IsComparable(params moqTypeCache_IsComparable_params) string {
-	return fmt.Sprintf("IsComparable(%#v)", params.expr)
+	return fmt.Sprintf("IsComparable(%#v, %#v)", params.expr, params.parentType)
 }
 
 func (m *moqTypeCache) paramsKey_IsComparable(params moqTypeCache_IsComparable_params, anyParams uint64) moqTypeCache_IsComparable_paramsKey {
@@ -1706,20 +1742,38 @@ func (m *moqTypeCache) paramsKey_IsComparable(params moqTypeCache_IsComparable_p
 			exprUsedHash = hash.DeepHash(params.expr)
 		}
 	}
+	var parentTypeUsed ast.TypeInfo
+	var parentTypeUsedHash hash.Hash
+	if anyParams&(1<<1) == 0 {
+		if m.runtime.parameterIndexing.IsComparable.parentType == moq.ParamIndexByValue {
+			parentTypeUsed = params.parentType
+		} else {
+			parentTypeUsedHash = hash.DeepHash(params.parentType)
+		}
+	}
 	return moqTypeCache_IsComparable_paramsKey{
-		params: struct{ expr dst.Expr }{
-			expr: exprUsed,
+		params: struct {
+			expr       dst.Expr
+			parentType ast.TypeInfo
+		}{
+			expr:       exprUsed,
+			parentType: parentTypeUsed,
 		},
-		hashes: struct{ expr hash.Hash }{
-			expr: exprUsedHash,
+		hashes: struct {
+			expr       hash.Hash
+			parentType hash.Hash
+		}{
+			expr:       exprUsedHash,
+			parentType: parentTypeUsedHash,
 		},
 	}
 }
 
-func (m *moqTypeCache_recorder) IsDefaultComparable(expr dst.Expr) *moqTypeCache_IsDefaultComparable_fnRecorder {
+func (m *moqTypeCache_recorder) IsDefaultComparable(expr dst.Expr, parentType ast.TypeInfo) *moqTypeCache_IsDefaultComparable_fnRecorder {
 	return &moqTypeCache_IsDefaultComparable_fnRecorder{
 		params: moqTypeCache_IsDefaultComparable_params{
-			expr: expr,
+			expr:       expr,
+			parentType: parentType,
 		},
 		sequence: m.moq.config.Sequence == moq.SeqDefaultOn,
 		moq:      m.moq,
@@ -1737,6 +1791,11 @@ func (r *moqTypeCache_IsDefaultComparable_fnRecorder) any() *moqTypeCache_IsDefa
 
 func (a *moqTypeCache_IsDefaultComparable_anyParams) expr() *moqTypeCache_IsDefaultComparable_fnRecorder {
 	a.recorder.anyParams |= 1 << 0
+	return a.recorder
+}
+
+func (a *moqTypeCache_IsDefaultComparable_anyParams) parentType() *moqTypeCache_IsDefaultComparable_fnRecorder {
+	a.recorder.anyParams |= 1 << 1
 	return a.recorder
 }
 
@@ -1900,7 +1959,7 @@ func (r *moqTypeCache_IsDefaultComparable_fnRecorder) repeat(repeaters ...moq.Re
 }
 
 func (m *moqTypeCache) prettyParams_IsDefaultComparable(params moqTypeCache_IsDefaultComparable_params) string {
-	return fmt.Sprintf("IsDefaultComparable(%#v)", params.expr)
+	return fmt.Sprintf("IsDefaultComparable(%#v, %#v)", params.expr, params.parentType)
 }
 
 func (m *moqTypeCache) paramsKey_IsDefaultComparable(params moqTypeCache_IsDefaultComparable_params, anyParams uint64) moqTypeCache_IsDefaultComparable_paramsKey {
@@ -1914,12 +1973,29 @@ func (m *moqTypeCache) paramsKey_IsDefaultComparable(params moqTypeCache_IsDefau
 			exprUsedHash = hash.DeepHash(params.expr)
 		}
 	}
+	var parentTypeUsed ast.TypeInfo
+	var parentTypeUsedHash hash.Hash
+	if anyParams&(1<<1) == 0 {
+		if m.runtime.parameterIndexing.IsDefaultComparable.parentType == moq.ParamIndexByValue {
+			parentTypeUsed = params.parentType
+		} else {
+			parentTypeUsedHash = hash.DeepHash(params.parentType)
+		}
+	}
 	return moqTypeCache_IsDefaultComparable_paramsKey{
-		params: struct{ expr dst.Expr }{
-			expr: exprUsed,
+		params: struct {
+			expr       dst.Expr
+			parentType ast.TypeInfo
+		}{
+			expr:       exprUsed,
+			parentType: parentTypeUsed,
 		},
-		hashes: struct{ expr hash.Hash }{
-			expr: exprUsedHash,
+		hashes: struct {
+			expr       hash.Hash
+			parentType hash.Hash
+		}{
+			expr:       exprUsedHash,
+			parentType: parentTypeUsedHash,
 		},
 	}
 }
